Support up to 2^32 bins

+++ include/seqan/kmer/filtervector.h
@@ -53,10 +53,10 @@ struct FilterVector<Uncompressed>
     static const uint16_t FILTER_METADATA_SIZE{256};
     static const uint8_t INT_SIZE{0x40};
 
-    uint16_t noOfBins;
+    uint32_t noOfBins;
     uint64_t noOfBits;
     uint16_t binWidth;
-    uint16_t blockBitSize;
+    uint32_t blockBitSize;
     uint64_t noOfBlocks;
     uint8_t  noOfChunks;
     uint64_t chunkSize;
@@ -73,7 +73,7 @@ struct FilterVector<Uncompressed>
 
     FilterVector() {}
 
-    FilterVector(uint16_t bins, uint64_t bits):
+    FilterVector(uint32_t bins, uint64_t bits):
         noOfBins(bins),
         noOfBits(bits)
     {
@@ -181,10 +181,10 @@ struct FilterVector<CompressedSimple>
     static const uint16_t FILTER_METADATA_SIZE{256};
     static const uint8_t INT_SIZE{0x40};
 
-    uint16_t noOfBins;
+    uint32_t noOfBins;
     uint64_t noOfBits;
     uint16_t binWidth;
-    uint16_t blockBitSize;
+    uint32_t blockBitSize;
     uint64_t noOfBlocks;
     uint8_t  noOfChunks;
     uint64_t chunkSize;
@@ -244,7 +244,7 @@ struct FilterVector<CompressedSimple>
 
     FilterVector() {}
 
-    FilterVector(uint16_t bins, uint64_t bits):
+    FilterVector(uint32_t bins, uint64_t bits):
         noOfBins(bins),
         noOfBits(bits)
     {
@@ -366,10 +366,10 @@ struct FilterVector<CompressedArray>
     static const uint8_t INT_SIZE{0x40};
     static const uint64_t MAX_VEC = 1ULL<<32; //512 MB, 36 -> 8GB, 39 -> 64
 
-    uint16_t noOfBins;
+    uint32_t noOfBins;
     uint64_t noOfBits;
     uint16_t binWidth;
-    uint16_t blockBitSize;
+    uint32_t blockBitSize;
     uint64_t noOfBlocks;
     uint8_t  noOfChunks;
     uint64_t chunkSize;
@@ -422,7 +422,7 @@ struct FilterVector<CompressedArray>
 
     FilterVector() {}
 
-    FilterVector(uint16_t bins, uint64_t bits):
+    FilterVector(uint32_t bins, uint64_t bits):
         noOfBins(bins),
         noOfBits(bits)
     {
